requirements.txt
torch>=2.0.0
lightning>=2.0.0
pybullet>=3.2.5
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
pillow>=9.5.0
imageio>=2.31.0
